requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.15.0
pandas>=2.0
pyarrow>=12.0.0
python-dotenv>=1.0.0
requests>=2.25.0